arrow_cast::cast::decimal

Function cast_decimal_to_float

pub(crate) fn cast_decimal_to_float<D: DecimalType, T: ArrowPrimitiveType, F>(
    array: &dyn Array,
    op: F,
) -> Result<ArrayRef, ArrowError>
where
    F: Fn(D::Native) -> T::Native,

Cast a decimal array to a floating point array.

Conversion is lossy and follows standard floating point semantics. Values that exceed the representable range become INFINITY or -INFINITY without returning an error.