arrow_cast::cast

Module list

Functions

cast_fixed_size_list_to_list 🔒

Cast fixed size list arrays to list arrays, maintaining the lengths of the inner lists.

cast_fixed_size_list_to_list_inner 🔒

cast_fixed_size_list_to_list_view 🔒

Same as cast_fixed_size_list_to_list but output list view array.

cast_list 🔒

Casting between list arrays of different offset size (e.g. List -> LargeList)

cast_list_to_fixed_size_list 🔒

Cast list to fixed size list array. If any inner list size does not match the size of the output fixed size list array, depending on cast_options we either output NULL for that element (safe) or raise an error.

cast_list_to_list_view 🔒

Casting list arrays to list view.

cast_list_values 🔒

Casting between list arrays of same offset size; we cast only the inner type.

cast_list_view 🔒

Casting between list view arrays of different offset size (e.g. ListView -> LargeListView)

cast_list_view_to_fixed_size_list 🔒

Same as cast_list_to_fixed_size_list but for list view arrays.

cast_list_view_to_list 🔒

Casting list view arrays to list.

cast_list_view_values 🔒

Casting between list view arrays of same offset size; we cast only the inner type.

cast_single_element_fixed_size_list_to_values 🔒

Cast fixed size list array to inner values type, essentially flattening the lists.

cast_values_to_fixed_size_list 🔒

Same as cast_values_to_list but output fixed size list array with element size 1.

cast_values_to_list 🔒

Converts a non-list array to a list array where every element is a single element list. NULLs in the original array become [NULL] (i.e. output list array contains no nulls since it wraps all input nulls in a single element list).

cast_values_to_list_view 🔒

Same as cast_values_to_list but output list view array.