arrow_cast::cast

Module map

Functions

cast_map_values 🔒

Helper function that takes a map container and casts the inner datatype.

key_field 🔒

Gets the key field from the entries of a map. For all other types returns None.

value_field 🔒

Gets the value field from the entries of a map. For all other types returns None.