arrow_cast

Module base64

Functions for converting data in [GenericBinaryArray] such as StringArray to/from base64 encoded strings

Constants

BASE64_STANDARD

A [GeneralPurpose] engine using the [alphabet::STANDARD] base64 alphabet and [PAD] config.

BASE64_STANDARD_NO_PAD

A [GeneralPurpose] engine using the [alphabet::STANDARD] base64 alphabet and [NO_PAD] config.

BASE64_URL_SAFE

A [GeneralPurpose] engine using the [alphabet::URL_SAFE] base64 alphabet and [PAD] config.

BASE64_URL_SAFE_NO_PAD

A [GeneralPurpose] engine using the [alphabet::URL_SAFE] base64 alphabet and [NO_PAD] config.

Traits

Engine

An Engine provides low-level encoding and decoding operations that all other higher-level parts of the API use. Users of the library will generally not need to implement this.

Functions

b64_decode

Base64 decode each element of array with the provided Engine

b64_encode

Bas64 encode each element of array with the provided Engine