arrow_cast

Module cast

Cast kernels to convert [ArrayRef] between supported datatypes.

See cast_with_options for more information on specific conversions.

Example:

// int32 to float64
let a = Int32Array::from(vec![5, 6, 7]);
let b = cast(&a, &DataType::Float64).unwrap();
let c = b.as_primitive::<Float64Type>();
assert_eq!(5.0, c.value(0));
assert_eq!(6.0, c.value(1));
assert_eq!(7.0, c.value(2));

Modules

decimal 🔒

dictionary 🔒

list 🔒

map 🔒

string 🔒

Structs

CastOptions

CastOptions provides a way to override the default cast behaviors

Functions

adjust_timestamp_to_timezone 🔒

as_time_res_with_timezone 🔒

bool_to_numeric_cast 🔒

can_cast_types

Return true if a value of type from_type can be cast into a value of to_type.

cast

Cast array to the provided data type and return a new Array with type to_type, if possible.

cast_binary_to_fixed_size_binary 🔒

Helper function to cast from one BinaryArray or ‘LargeBinaryArray’ to ‘FixedSizeBinaryArray’.

cast_bool_to_numeric 🔒

Cast Boolean types to numeric

cast_byte_container 🔒

Helper function to cast from one ByteArrayType to another and vice versa. If the target one (e.g., LargeUtf8) is too large for the source array it will return an Error.

cast_duration_to_interval 🔒

Cast the array from duration and interval

cast_fixed_size_binary_to_binary 🔒

Helper function to cast from ‘FixedSizeBinaryArray’ to one BinaryArray or ‘LargeBinaryArray’. If the target one is too large for the source array it will return an Error.

cast_from_decimal 🔒

cast_integer_to_decimal 🔒

cast_interval_day_time_to_interval_month_day_nano 🔒

Cast the array from interval day time to month day nano

cast_interval_year_month_to_interval_month_day_nano 🔒

Cast the array from interval year month to month day nano

cast_month_day_nano_to_duration 🔒

Cast the array from interval to duration

cast_numeric_arrays 🔒

Convert Array into a PrimitiveArray of type, and apply numeric cast

cast_numeric_to_binary 🔒

cast_numeric_to_bool 🔒

Cast numeric types to Boolean

cast_reinterpret_arrays 🔒

Cast the primitive array using [PrimitiveArray::reinterpret_cast]

cast_to_decimal 🔒

cast_view_to_byte 🔒

Helper function to cast from one ByteViewType array to ByteArrayType array.

cast_with_options

Try to cast array to to_type if possible.

make_duration_array 🔒

make_timestamp_array 🔒

numeric_cast 🔒

numeric_to_bool_cast 🔒

time_unit_multiple 🔒

Get the time unit as a multiple of a second

timestamp_to_date32 🔒

try_numeric_cast 🔒