arrow_cast::cast

Module decimal

Traits

DecimalCast ğŸ”’

A utility trait that provides checked conversions between decimal types inspired by NumCast

Functions

cast_decimal_to_decimal ğŸ”’

cast_decimal_to_decimal_error ğŸ”’

cast_decimal_to_decimal_same_type ğŸ”’

cast_decimal_to_float ğŸ”’

cast_decimal_to_integer ğŸ”’

cast_floating_point_to_decimal ğŸ”’

cast_string_to_decimal ğŸ”’

Cast Utf8 to decimal

convert_to_bigger_or_equal_scale_decimal ğŸ”’

convert_to_smaller_scale_decimal ğŸ”’

generic_string_to_decimal_cast ğŸ”’

parse_string_to_decimal_native ğŸ”’

Parses given string to specified decimal native (i128/i256) based on given scale. Returns an Err if it cannot parse given string.

string_to_decimal_cast ğŸ”’

string_view_to_decimal_cast ğŸ”’