Module arrow_cast::cast::decimal

Traits

DecimalCast 🔒

A utility trait that provides checked conversions between decimal types inspired by NumCast

Functions

cast_decimal_to_decimal 🔒

cast_decimal_to_decimal_error 🔒

cast_decimal_to_decimal_same_type 🔒

cast_decimal_to_float 🔒

cast_decimal_to_integer 🔒

cast_floating_point_to_decimal128 🔒

cast_floating_point_to_decimal256 🔒

cast_string_to_decimal 🔒

Cast Utf8 to decimal

convert_to_bigger_or_equal_scale_decimal 🔒

convert_to_smaller_scale_decimal 🔒

parse_string_to_decimal_native 🔒

Parses given string to specified decimal native (i128/i256) based on given scale. Returns an Err if it cannot parse given string.

string_to_decimal_cast 🔒