arrow_cast::cast
Module list

Functions

cast_fixed_size_list_to_list ğŸ”’

cast_list ğŸ”’

Cast the container type of List/Largelist array along with the inner datatype

cast_list_to_fixed_size_list ğŸ”’

cast_list_values ğŸ”’

Helper function that takes an Generic list container and casts the inner datatype.

cast_single_element_fixed_size_list_to_values ğŸ”’

cast_values_to_fixed_size_list ğŸ”’

Helper function that takes a primitive array and casts to a fixed size list array.

cast_values_to_list ğŸ”’

Helper function that takes a primitive array and casts to a (generic) list array.