arrow_cast

Module display

Functions for printing array values as human-readable strings.

This is often used for debugging or logging purposes.

See the pretty crate for additional functions for record batch pretty printing.

Macros

decimal_display 🔒

duration_display 🔒

Derive DisplayIndexState for PrimitiveArray<$t>

duration_fmt 🔒

duration_option_display 🔒

Similar to duration_display but $convert returns an Option

primitive_display 🔒

primitive_display_float 🔒

temporal_display 🔒

timestamp_display 🔒

Structs

ArrayFormat 🔒

ArrayFormatter

A string formatter for an [Array]

FormatOptions

Options for formatting arrays

MillisecondsFormatter 🔒

NanosecondsFormatter 🔒

ValueFormatter

Implements Display for a specific array value

Enums

DurationFormat

Format for displaying durations

FormatError

Either an [ArrowError] or std::fmt::Error

Traits

ArrayFormatterFactory

Allows creating a new ArrayFormatter for a given [Array] and an optional [Field].

DisplayIndex

Display but accepting an index

DisplayIndexState 🔒

DisplayIndex with additional state

Functions

array_format 🔒

array_value_to_string

Get the value at the given row in an array as a String.

date32_to_date 🔒

lexical_to_string

Converts numeric type to a String

make_array_formatter 🔒

Used to create a new ArrayFormatter from the given array, while also checking whether there is an override available in the ArrayFormatterFactory.

make_default_display_index 🔒

write_list 🔒

write_timestamp 🔒

Type Aliases

FieldDisplay 🔒

Pairs an ArrayFormatter with its field name

FormatResult

The result of formatting an array element via DisplayIndex::write.

TimeFormat 🔒