arrow_cast

Module display

Functions for printing array values as human-readable strings.

This is often used for debugging or logging purposes.

See the pretty crate for additional functions for record batch pretty printing.

Macros

decimal_display ğŸ”’

duration_display ğŸ”’

duration_fmt ğŸ”’

primitive_display ğŸ”’

primitive_display_float ğŸ”’

temporal_display ğŸ”’

timestamp_display ğŸ”’

Structs

ArrayFormat ğŸ”’

ArrayFormatter

A string formatter for an [Array]

FormatOptions

Options for formatting arrays

MillisecondsFormatter ğŸ”’

NanosecondsFormatter ğŸ”’

ValueFormatter

Implements Display for a specific array value

Enums

DurationFormat

Format for displaying durations

FormatError ğŸ”’

Either an [ArrowError] or std::fmt::Error

Traits

DisplayIndex ğŸ”’

Display but accepting an index

DisplayIndexState ğŸ”’

DisplayIndex with additional state

Functions

array_format ğŸ”’

array_value_to_string

Get the value at the given row in an array as a String.

date32_to_date ğŸ”’

lexical_to_string

Converts numeric type to a String

make_formatter ğŸ”’

write_list ğŸ”’

write_timestamp ğŸ”’

Type Aliases

FieldDisplay ğŸ”’

Pairs a boxed DisplayIndex with its field name

FormatResult ğŸ”’

TimeFormat ğŸ”’