Module arrow_cast::parse

Traits

Parser

Specialized parsing implementations used by csv and json reader

Functions

parse_decimal

Parse the string format decimal value to i128/i256 format and checking the precision and scale. The result value can’t be out of bounds.

parse_interval_day_time

parse_interval_month_day_nano

parse_interval_year_month

string_to_datetime

Accepts a string and parses it relative to the provided timezone

string_to_time_nanoseconds

Accepts a string in ISO8601 standard format and some variants and converts it to nanoseconds since midnight.

string_to_timestamp_nanos

Accepts a string in RFC3339 / ISO8601 standard format and some variants and converts it to a nanosecond precision timestamp.

Type Aliases

MonthDayNano