Module arrow_cast::parse

Macros

parser_primitive 🔒

Structs

Interval 🔒

IntervalAmount 🔒

IntervalParseConfig 🔒

TimestampParser 🔒

Helper for parsing RFC3339 timestamps

Enums

IntervalUnit 🔒

Constants

EPOCH_DAYS_FROM_CE 🔒

Number of days between 0001-01-01 and 1970-01-01

ERR_NANOSECONDS_NOT_SUPPORTED 🔒

Error message if nanosecond conversion request beyond supported interval

INTERVAL_PRECISION 🔒

Chosen based on the number of decimal digits in 1 week in nanoseconds

NANOS_PER_HOUR 🔒

NANOS_PER_MILLIS 🔒

NANOS_PER_MINUTE 🔒

NANOS_PER_SECOND 🔒

Traits

Parser

Specialized parsing implementations used by csv and json reader

Functions

parse_date 🔒

parse_decimal

Parse the string format decimal value to i128/i256 format and checking the precision and scale. The result value can’t be out of bounds.

parse_interval_components 🔒

parse the string into a vector of interval components i.e. (amount, unit) tuples

parse_interval_day_time

parse_interval_month_day_nano

parse_interval_year_month

parse_nanos 🔒

Parse nanoseconds from the first N values in digits, subtracting the offset O

string_to_datetime

Accepts a string and parses it relative to the provided timezone

string_to_time 🔒

string_to_time_nanoseconds

Accepts a string in ISO8601 standard format and some variants and converts it to nanoseconds since midnight.

string_to_timestamp_nanos

Accepts a string in RFC3339 / ISO8601 standard format and some variants and converts it to a nanosecond precision timestamp.

to_timestamp_nanos 🔒

Fallible conversion of NaiveDateTime to i64 nanoseconds

Type Aliases

MonthDayNano