arrow_cast
Module pretty

Utilities for pretty printing RecordBatches and Arrays.

Note this module is not available unless feature = "prettyprint" is enabled.

Functions

create_column 🔒

create_table 🔒

Convert a series of record batches into a table

pretty_format_batches

Create a visual representation of record batches

pretty_format_batches_with_options

Create a visual representation of record batches

pretty_format_columns

Create a visual representation of columns

pretty_format_columns_with_options 🔒

Utility function to create a visual representation of columns with options

print_batches

Prints a visual representation of record batches to stdout

print_columns

Prints a visual representation of a list of column to stdout